+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+# Unreleased
+- custom offset grammar: `alos:18.5deg`, `shaah_zmanis_mga:80min`, `tzeis:100minz`
+
 # 0.1.1
 - nushell completions
 - *zman* name completions for bash, zsh & nushell